requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pydantic[email]
motor==2.5.1
bcrypt==3.2.0
passlib
phonenumbers==8.12.23
PyJWT==2.3.0
pymongo==3.12.0
python-multipart==0.0.5
pytest
mock==4.0.3
pytest-asyncio==0.16.0
python-decouple==3.4
pytest-mock
pytest-cov